LinearEphemeris.jl

Documentation for LinearEphemeris.jl

Package for fitting linear ephemeris to a series of transit times (T0s) of an exoplanet.

1. Introduction

A Julia package for linear ephemerides fit of exoplanet transit times. It uses the Weighted Least Square (WLS) fit based on the Numerical Recipes. It can evaluate the errors on the reference time (Tref, that is the intercept) and on the period (Pref, that is the slope) based on formal error of the WLS or with a bootstrap approach. It returns to the standard output a summary of the fit (with statistics) and each transit time (T0) with own corresponding epoch (or transit number), expected transit time (Tlin) from the linear ephem. and O-C (Observed - Calculated, where O == T0 and C == Tlin), and the source (that is the telescope used to observe that transit).

2. Install

julia> using Pkg
julia> Pkg.clone("https://github.com/lucaborsato/LinearEphemeris.jl.git")

3. Example

Create/Read list of T0s = [1234.9034, ..., 16789.3980], errors err_T0s = [0.00034, ..., 0.00045]. Define input Tref, Pref. Using default parameters, run as:

using LinearEphemeris
full_linear_ephemeris_analysis(
    T0s,
    err_T0s;
    Tref_in = Tref,
    Pref_in = Pref
)

It will output to screen a summary of the fit with statistics, bootstrap error analysis, and it will open a plot gui.

You can provide also a list of `sources = ["TESS", ..., "K2"] and it will plot each telescope/source with a different color.

The following is a call of the function with the allowed parameters and default values:

full_linear_ephemeris_analysis(
        T0s,
        err_T0s;
        Tref_in = nothing,
        Pref_in = nothing,
        tscale = nothing,
        sources = nothing,
        bootstrap = true,
        nboot = nothing,
        do_plot = true,
        show_gui = true,
        plot_file = nothing,
        seed = 42,
    )

where: tscale is a parameter to be subtracted to the x-axis, only for plot. nboot is the number of bootstrap iterations, if nothing it will computed automatically as n (log(n))^2, with n the number of T0s. do_plot a keyword to do the plot or not. plot_file is the file name (with complete path) of the output of the plot, if do_plot = false it will not be used. seed is the seed for the random number generator.

There is a working example based on Kepler- 9 in the examples folder, just run Kepler-9.jl and it will print out the results and create the two O-C diagrams (one for each planet).

4. Function Documentation

LinearEphemeris.linear_error_prop — Function

linear_error_prop(eq, em, x)

propagates error on x value as:

\[\sqrt{\mathrm{eq}^2 + (x × \mathrm{em})^2}\]

Arguments:

eq::Float64: error on intercept (reference time Tref)

em::Float64: error on slope (reference period Pref)

x::Float64: value at which compute the error

LinearEphemeris.transit_epoch — Function

transit_epoch(T0, Tref, Pref)

computes the transit epoch (or transit number N) with respect to a linear ephemeris:

\[ N = \mathrm{round}( (T_{0,\mathrm{lin}} - T_\mathrm{ref}) / P_\mathrm{ref}) \]

LinearEphemeris.linear_transit_time — Function

linear_transit_time(Tref, Pref, epoch)

computes the predicted transit time at specific epoch (transit number N) given a linear ephemeris:

\[ T_{0,\mathrm{lin}, N} = T_\mathrm{ref} + N \times P_\mathrm{ref} \]

LinearEphemeris.linear_transit_times — Function

linear_transit_times(Tref, Pref, epochs)

as linear_transit_time but for vector/list of epochs

LinearEphemeris.wls_nr — Function

wls_nr(x, y, ey)

WEIGHTED LEAST SQUARE AS IN NUMERICAL RECIPES

Arguments:

x::Vector{Float64}: independent variable

y::Vector{Float64}: dependent variable

ey::Vector{Float64}: measurement errors on dependent variable

Returns:

intercept::tuple(fit::Float64, err::Float64): intercept with value (fit) and formal error (err)

slope::tuple(fit::Float64, err::Float64): slope with value (fit) and formal error (err)

LinearEphemeris.fit_linear_ephemeris — Function

fit_linear_ephemeris(epochs, T0s, err_T0s)

as in wls_nr but with meaningful variable names

Arguments:

epochs::Vector{Float64}: transit epochs/numbers

T0s::Vector{Float64}: transit times

err_T0s::Vector{Float64}: measurement errors on T0s

Returns:

Tref::Tuple(fit::Float64, err::Float64): reference time of the linear ephemeris with value (fit) and formal error (err)

Pref::Tuple(fit::Float64, err::Float64): period of the linear ephemeris with value (fit) and formal error (err)

LinearEphemeris.statistics_fit — Function

statistics_fit(y_obs, y_mod; ey_obs = nothing, n_fit = 2)

computes some statistics on the fitted linear ephemeris

Arguments:

y_obs::Vector{Float64} observed values (T0s)

y_mod::Vector{Float64} model values (T0s,lin)

ey_obs::Vector{Float64} = nothing errors on the observed values (if available)

n_fit::Int = 2 number of fitted parameters of the model (2 for the linear ephemeris)

Returns:

stats::NamedTuple: named tuple with various statistics based on the residuals

stats.std::Float64 standard deviation of the residuals

stats.p68::Float64 68.27th percentile of the absolute residuals

stats.chi_square::Float64 χ^2 like = ∑((obs - mod) / err_obs)^2

stats.red_chi_square::Float64 χ^2 / dof

stats.dof::Float64 degrees of freedom dof = n_data-n_fit

stats.bic::Float64 BIC= χ^2 + n_fit ln(n_data)

stats.aic::Float64 AIC= χ^2 + 2 n_fit

stats.lnc::Float64 log-Likelihood constant part given by the 2π and σ_i terms

stats.lnL::Float64 log-Likelihood

stats.bic_lnL::Float64 BIC= -2lnL + n_fit ln(n_data)

stats.aic_lnL::Float64 AIC= -2lnL + 2 n_fit

LinearEphemeris.get_nboot — Function

get_nboot(n)

Computes the minimum proposed number of bootstrap iteration as:

\[ n_\mathrm{boot} = n \times \mathrm{round}(Int, \ln(n)^2)\]

Arguments:

n::Int number of observed transit times (T0s)

Returns:

n_boot::Int number of bootstrap iterations

LinearEphemeris.classical_bootstrap — Function

classical_bootstrap(
    x,
    y, 
    q, 
    m; 
    ey_in = nothing, 
    nboot = nothing, 
    rng=Random.GLOBAL_RNG,
    return_distribution=true
)

Arguments:

x::Vector{Float64}: independent variable

y::Vector{Float64}: dependent variable

q::Float: intercept (the true or best-fit value)

m::Float: slope (the true or best-fit value)

ey_in::Vector{Float64} = nothing: measurement errors on dependent variable (default is nothing)

nboot::Int = nothing: number of bootstrap iterations to run

rng = Random.GLOBAL_RNG: random number generator

return_distribution::Bool = true: return or not the bootstrap distribution of each parameter

Return:

tuple(median::Float64, rms=Float64, distribution=Vector{Float64}): intercept bootstrap (median, rms, distribution); return_distribution=false will set distribution=nothing

tuple(median::Float64, rms=Float64, distribution=Vector{Float64}): slop bootstrap

LinearEphemeris.full_linear_ephemeris_analysis — Function

full_linear_ephemeris_analysis(
    T0s,
    err_T0s;
    Tref_in = nothing,
    Pref_in = nothing,
    tscale = nothing,
    sources = nothing,
    bootstrap = true,
    nboot = nothing,
    do_plot = true,
    show_gui = true,
    plot_file = nothing,
    seed = 42,
    return_distribution = true,
)

Compute the full linear ephemeris with all needed arguments. It prints the output on the standard output and do the plots if required.

Arguments:

T0s::Vector{Float64}: observed transit times vector

err_T0s::Vector{Float64}: measurement error vector on observed transit times

Tref_in::Float64 = nothing: input reference time for the linear ephemeris

Pref_in::Float64 = nothing: input reference period for the linear ephemeris

tscale::Float64 = nothing: scale time, just for plot

sources::Vector{String} = nothing: list/vector of sources of each transit time

bootstrap::Bool = true: run or not the bootstrap analysis

nboot::Int = nothing: number of bootstrap iterations if required

return_distribution = true: in case of bootstrap, return distribution (if also do_plot, it plots distribution)

do_plot::Bool = true: do or not the plot

show_gui::Bool = true: if do_plot, this flag open the GUI or not

plot_file::String = nothing: if do_plot, this is the full name of the output plot file

seed::Int = 42: seed of the random number

Returns:

nothing

There are some more functions in the package, but I did not commented them because they are like sub-small pieces of others functions.

Index

LinearEphemeris.classical_bootstrap

LinearEphemeris.fit_linear_ephemeris

LinearEphemeris.full_linear_ephemeris_analysis

LinearEphemeris.get_nboot

LinearEphemeris.linear_error_prop

LinearEphemeris.linear_transit_time

LinearEphemeris.linear_transit_times

LinearEphemeris.statistics_fit

LinearEphemeris.transit_epoch

LinearEphemeris.wls_nr